third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-tabbed-editor-opens-network-uisourcecode-expected.txt

Verify that for a fileSystem UISourceCode with network binding TabbedEditorContainer opens network UISourceCode.

Binding created: {
       network: http://127.0.0.1:8000/inspector/persistence/resources/foo.js
    fileSystem: file:///var/www/inspector/persistence/resources/foo.js
}
request open uiSourceCode: file:///var/www/inspector/persistence/resources/foo.js
Opened tabs: 
    http://127.0.0.1:8000/inspector/persistence/resources/foo.js
